+++ packages/e2e/src/integrations/cli/createMigration/testConfig.ts
@@ -1,5 +1,5 @@
 import * as path from 'path';
-import { Config } from '@elasticsearch-heaven/types';
+import { Config } from '@tshio/elasticsearch-heaven-types';
 
 export default {
   clientOptions: {