+++ packages/elasticsearch-heaven-core/jest.config.js
@@ -1,0 +1,17 @@
+module.exports = {
+  displayName: 'elasticsearch-heaven-core',
+  preset: '../../jest.preset.js',
+  globals: {
+    'ts-jest': {
+      tsconfig: '<rootDir>/tsconfig.spec.json',
+    },
+  },
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.[tj]sx?$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory: '../../coverage/packages/elasticsearch-heaven-core',
+  setupFilesAfterEnv: ['../../tests/setup.ts'],
+  testTimeout: 500000,
+};